fix(basler+detect): review findings — UAF, stagger, plate consensus, direction
Correctness/robustness fixes from the module review:

Basler:
- HIGH: manual override is now applied to the device BEFORE it is published to the
  shared maps (removing reapplyManualIfAny), so the reconnect worker no longer
  mutates an already-published exposure strategy / camera nodes that the capture
  thread is using — closes a use-after-free + concurrent Pylon-node race.
- Stagger index is bounded by the per-NIC camera count, not the global total, so
  two cameras on one NIC can't get an identical GevSCFTD offset.
- applyRoi clamps aligned offsets up to the node minimum (a sub-min offset no
  longer throws).
- CameraDevice caches PixelFormat colorness and guards GetValue with IsReadable,
  so a transient unreadable format doesn't force a reconnect.

Direction:
- The confirmSightings backstop now finalises the STANDING (reported) decision,
  not the current frame — a single late outlier can no longer win the lock.
- Bounded the size/y deques (idling vehicle can't grow them unboundedly).
- passIdFor no longer erases the reference plate on an empty (no-OCR) frame.
- DetectionWorker passGapMs is kept in sync with the estimator trackGapMs so a
  quiet gap resets both together; removed the dead dirByText map.

Plate accuracy:
- PlateProcessor emits the accuracy-WEIGHTED consensus text (highest cumulative
  confidence across the pass's reads) instead of the single highest-confidence
  frame, so one over-confident misread can't override an agreeing majority.

Live-adapt millisecond counters use 64-bit (long long).

+++ include/lpr/basler/GigeRateAdapt.h
@@ -17,7 +17,7 @@ struct GigeAdaptCfg {
     double downF     = 0.75;   // throttle-down step multiplier
     double upF       = 1.15;   // recovery step multiplier
     double minFps    = 1.0;    // floor
-    long   recoverMs = 15000;  // clean time before a recovery step
+    long long recoverMs = 15000;  // clean time before a recovery step
 };
 
 enum class GigeAdaptAction { None, Throttle, Recover };
@@ -29,8 +29,8 @@ struct GigeAdaptDecision {
 
 // `cur` = current fps, `ceil` = connect-time cap, `lastStepMs` = when the last step
 // was applied, `lossPct` = measured loss over the window, `nowMs` = steady clock ms.
-inline GigeAdaptDecision gigeAdaptDecide(double cur, double ceil, long lastStepMs,
-                                         double lossPct, long nowMs, const GigeAdaptCfg& c) {
+inline GigeAdaptDecision gigeAdaptDecide(double cur, double ceil, long long lastStepMs,
+                                         double lossPct, long long nowMs, const GigeAdaptCfg& c) {
     if (lossPct >= c.downPct && cur > c.minFps)
         return { GigeAdaptAction::Throttle, std::max(c.minFps, cur * c.downF) };
     if (lossPct <= c.upPct && cur < ceil && (nowMs - lastStepMs) >= c.recoverMs)
